# docs/api.yaml
openapi: 3.0.3
info:
  title: EcoTn API
  version: 0.0.1
  description: API REST per la gestione di persone, abitazioni, consumi, produzione, notizie e chiavi.
servers:
  - url: http://localhost:3000
    description: Server locale di sviluppo
paths:
...
  /api/v1/houses:
    get:
      summary: Lista di tutte le abitazioni
      description: Restituisce l'elenco completo di tutte le abitazioni, con i dati delle persone associate popolati.
      operationId: getHouses
      responses:
        "200":
          description: Elenco di abitazioni
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/House"
        "500":
          description: Errore interno del server
    post:
      summary: Crea una nuova abitazione
      description: Registra una nuova abitazione nel sistema a partire dai dati forniti nel body.
      operationId: createHouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/House"
      responses:
        "201":
          description: Abitazione creata con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
        "400":
          description: Dati non validi
  /api/v1/houses/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID dell'abitazione (ObjectId)
        schema:
          type: string
    get:
      summary: Ottiene un'abitazione tramite ID
      description: Restituisce i dati di una singola abitazione identificata dal suo ObjectId, con le persone associate popolate.
      operationId: getHouseById
      responses:
        "200":
          description: Abitazione trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
        "404":
          description: Abitazione non trovata
        "500":
          description: Errore interno del server
    put:
      summary: Aggiorna un'abitazione esistente
      description: Sovrascrive i dati di un'abitazione esistente con quelli forniti nel body.
      operationId: updateHouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/House"
      responses:
        "200":
          description: Abitazione aggiornata con successo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/House"
        "400":
          description: Dati non validi
        "404":
          description: Abitazione non trovata
    delete:
      summary: Elimina un'abitazione
      description: Rimuove definitivamente dal sistema l'abitazione con l'ID specificato.
      operationId: deleteHouse
      responses:
        "200":
          description: Abitazione eliminata con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "House deleted"
        "404":
          description: Abitazione non trovata
        "500":
          description: Errore interno del server
  /api/v1/houses/{id}/consumption-stats:
    parameters:
      - name: id
        in: path
        required: true
        description: ID dell'abitazione (ObjectId)
        schema:
          type: string
      - name: start
        in: query
        required: true
        description: "Data di inizio del periodo (formato: YYYY-MM-DD). Se uguale a end, il raggruppamento è per ora."
        schema:
          type: string
          format: date
          example: "2024-01-01"
      - name: end
        in: query
        required: true
        description: "Data di fine del periodo (formato: YYYY-MM-DD). Se uguale a start, il raggruppamento è per ora."
        schema:
          type: string
          format: date
          example: "2024-01-31"
    get:
      summary: Statistiche di consumo di un'abitazione
      description: >
        Restituisce la media dei valori di consumo nell'intervallo indicato.
        Il raggruppamento è per giorno se start ≠ end, per ora se start = end.
      operationId: getHouseConsumptionStats
      responses:
        "200":
          description: Statistiche di consumo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        "500":
          description: Errore interno del server
  /api/v1/houses/{id}/production-stats:
    parameters:
      - name: id
        in: path
        required: true
        description: ID dell'abitazione (ObjectId)
        schema:
          type: string
      - name: start
        in: query
        required: true
        description: "Data di inizio del periodo (formato: YYYY-MM-DD). Se uguale a end, il raggruppamento è per ora."
        schema:
          type: string
          format: date
          example: "2024-01-01"
      - name: end
        in: query
        required: true
        description: "Data di fine del periodo (formato: YYYY-MM-DD). Se uguale a start, il raggruppamento è per ora."
        schema:
          type: string
          format: date
          example: "2024-01-31"
    get:
      summary: Statistiche di produzione di un'abitazione
      description: >
        Restituisce la media dei valori di produzione nell'intervallo indicato.
        Il raggruppamento è per giorno se start ≠ end, per ora se start = end.
      operationId: getHouseProductionStats
      responses:
        "200":
          description: Statistiche di produzione
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        "500":
          description: Errore interno del server
...
components:
  schemas:
...
    House:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId dell'abitazione
        indirizzo:
          type: string
          description: Indirizzo dell'abitazione
        immagineProfilo:
          type: string
          description: URL o path dell'immagine di profilo dell'abitazione
        persone:
          type: array
          description: Elenco degli ObjectId delle persone associate all'abitazione
          items:
            type: string
            description: ObjectId di una persona
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - indirizzo
        - persone
...
    Stats:
      type: object
      description: >
        Risultato aggregato delle statistiche di consumo o produzione.
        Le label sono date (YYYY-MM-DD) per range multi-giorno, oppure ore intere (0–23) per singolo giorno.
      properties:
        labels:
          type: array
          description: Etichette dell'asse X — date o ore a seconda dell'intervallo richiesto
          items:
            type: string
          example: ["2024-01-01", "2024-01-02", "2024-01-03"]
        values:
          type: array
          description: Valori medi corrispondenti a ciascuna label
          items:
            type: number
          example: [3.2, 4.1, 2.8]
...